+++ web/static/js/fm_fact_label/factlabel_visualization.js
@@ -176,9 +176,20 @@ function drawFMFactLabel(data) {
       var domainSize = domain.node().getBBox();
    }
 
+   // Language level
+   var languageLevelHeight = domainHeight + domainSize.height;
+   const langProp = get_property(data, 'Language level');
+   if (langProp == null || langProp.value === null) {
+      var languageLevelSize = { "width": 0, "height": 0 };
+   } else {
+      var languageLevel = chart.append("g").attr("transform", "translate(0," + languageLevelHeight + ")");
+      addMetadata(languageLevel, "Language level:", langProp.value);
+      var languageLevelSize = languageLevel.node().getBBox();
+   }
+
    // Reference
    if (!get_property(data, 'Reference').value == "") {
-      var reference = chart.append("g").attr("transform", "translate(0," + (domainHeight + domainSize.height - MAIN_RULE_HEIGHT - 10) + ")");
+      var reference = chart.append("g").attr("transform", "translate(0," + (languageLevelHeight + languageLevelSize.height - MAIN_RULE_HEIGHT - 10) + ")");
       reference.append('a')
          .attr("id", "hrefIcon")
          .attr("href", get_property(data, 'Reference').value)
@@ -194,7 +205,7 @@ function drawFMFactLabel(data) {
    }
 
    // Middle rule 1
-   yRule1 = domainHeight + domainSize.height;
+   yRule1 = languageLevelHeight + languageLevelSize.height;
    chart.append("g").attr("id", "rule1");
    drawRule("rule1", yRule1);